some how it did 873 present of profit I am amazingggggggg

+++ FinanceMaker.Common/Models/Pullers/YahooFinance/InterdayModel.cs
@@ -129,7 +129,7 @@
         public float?[] open { get; set; }
         public float?[] low { get; set; }
         public float?[] close { get; set; }
-        public int?[] volume { get; set; }
+        public long?[] volume { get; set; }
     }
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider adding the 'required' modifier or declaring as nullable.
